pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "bugagent"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = ["click>=8", "PyGithub>=2", "pyyaml>=6"]

[project.scripts]
bugagent = "bugagent.cli:main"
